feat: add tier to claim, move tier to common
