+++ common-tests/src/lib.rs
@@ -54,8 +54,12 @@ where
     }
 
     fn call(&mut self, mut req: hyper::Request<hyper::Body>) -> Self::Future {
-        req.extensions_mut()
-            .insert(Claim::new("test".to_string(), self.scopes.clone(), None));
+        req.extensions_mut().insert(Claim::new(
+            "test".to_string(),
+            self.scopes.clone(),
+            None,
+            None,
+        ));
         self.inner.call(req)
     }
 }
