+++ common/Cargo.toml
@@ -64,7 +64,8 @@ backend = [
     "tracing-subscriber/fmt",
     "ttl_cache",
     "sqlx",
-    "sqlx/postgres",                 # Fix for derive macro error when different backends enable sqlite & postgres (mainly for grouped compilation in clippy CI and Docker build)
+    "sqlx/postgres",            # Fix for derive macro error when different backends enable sqlite & postgres (mainly for grouped compilation in clippy CI and Docker build)
+    "sqlx/sqlite",
 ]
 claims = [
     "bytes",
